+++ CHANGELOG.md
@@ -5,12 +5,36 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.29.0] - 18-Aug-2024
+
+### Added
+
+- Initial worldspace UI support ([#304](https://github.com/mvlabat/bevy_egui/pull/304) by @TheButlah, @Schmarni-Dev)
+- Paint callback support ([#303](https://github.com/mvlabat/bevy_egui/pull/303) by @PPakalns)
+
+### Changed
+
+- Adapt to `web-sys` clipboard api change ([#301](https://github.com/mvlabat/bevy_egui/pull/301) by @no-materials).
+
+### Fixed
+
+- Clear modifier state when focus is lost ([#298](https://github.com/mvlabat/bevy_egui/pull/298) by @SludgePhD).
+- Fix redraws ([#293](https://github.com/mvlabat/bevy_egui/pull/293)).
+
+## [0.28.0] - 6-Jul-2024
+
+### Changed
+
+- Update Bevy to 0.14 ([#284](https://github.com/mvlabat/bevy_egui/pull/284) by @Friz64).
+- Update Egui to 0.28 ([#290](https://github.com/mvlabat/bevy_egui/pull/290) by @Swoorup).
+- Update webbrowser to 1.0.1
+
 ## [0.27.1] - 2-Jun-2024
 
 ### Changed
 
 - Request Redraw only if really needed ([#278](https://github.com/mvlabat/bevy_egui/pull/278) by @Maximetinu).
-- Fixed light in the `render_to_image_wideget` example ([#282](https://github.com/mvlabat/bevy_egui/pull/282) by @rlidwka).
+- Fix light in the `render_to_image_wideget` example ([#282](https://github.com/mvlabat/bevy_egui/pull/282) by @rlidwka).
 
 ## [0.27.0] - 18-Apr-2024